Add Utilities class with ensure_directory_exists method
