+++ pan_os_upgrade/components/utility.py
@@ -1,0 +1,67 @@
+import os
+
+
+class Utilities:
+    def __init__(self):
+        pass
+
+    # @staticmethod
+    # def compare_versions(version1: str, version2: str) -> str:
+    #     # Function implementation here
+
+    # @staticmethod
+    # def configure_logging(level: str, encoding: str = "utf-8", log_file_path: str = "logs/upgrade.log", log_max_size: int = 10 * 1024 * 1024) -> None:
+    #     # Function implementation here
+
+    # @staticmethod
+    # def get_emoji(action: str) -> str:
+    #     # Function implementation here
+
+    # @staticmethod
+    # def flatten_xml_to_dict(element: ET.Element) -> dict:
+    #     # Function implementation here
+
+    @staticmethod
+    def ensure_directory_exists(file_path: str) -> None:
+        """
+        Ensures the existence of the directory path for a given file path, creating it if necessary.
+
+        This function is crucial for file operations, particularly when writing to files, as it guarantees that the directory path exists prior to file creation or modification. It parses the provided file path to isolate the directory path and, if this directory does not exist, it creates it along with any required intermediate directories. This proactive approach prevents errors related to non-existent directories during file operations.
+
+        Parameters
+        ----------
+        file_path : str
+            The complete file path for which the existence of the directory structure is to be ensured. The function identifies the directory path component of this file path and focuses on verifying and potentially creating it.
+
+        Raises
+        ------
+        OSError
+            In the event of a failure to create the directory due to insufficient permissions or other filesystem-related errors, an OSError is raised detailing the issue encountered.
+
+        Examples
+        --------
+        Creating a directory structure for a log file:
+            >>> ensure_directory_exists('/var/log/my_application/error.log')
+            # This will check and create '/var/log/my_application/' if it does not already exist, ensuring a valid path for 'error.log'.
+
+        Notes
+        -----
+        - Employs `os.makedirs` with `exist_ok=True`, which allows the directory to be created without raising an exception if it already exists, ensuring idempotency.
+        - Designed to be platform-independent, thereby functioning consistently across various operating systems and Python environments, enhancing the function's utility across diverse application scenarios.
+        """
+
+        directory = os.path.dirname(file_path)
+        if not os.path.exists(directory):
+            os.makedirs(directory)
+
+    # @staticmethod
+    # def find_close_matches(available_versions: List[str], target_version: str, max_results: int = 5) -> List[str]:
+    #     # Function implementation here
+
+    # @staticmethod
+    # def parse_version(version: str) -> Tuple[int, int, int, int]:
+    #     # Function implementation here
+
+    # @staticmethod
+    # def resolve_hostname(hostname: str) -> bool:
+    #     # Function implementation here
